% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona.R
\name{as.SpatialPointsDataFrame}
\alias{as.SpatialPointsDataFrame}
\title{Produce a SpatialPointsDataFrame if data has a column of type `gps` or `geopoint`.
Otherwise, return NA.}
\usage{
as.SpatialPointsDataFrame(onaObj)
}
\arguments{
\item{the}{ona object which will be possibly co-erced to a SpatialPointsDataFrame object.}
}
\value{
A SpatialPointsDataFrame representation of this ona Object
}
\description{
Produce a SpatialPointsDataFrame if data has a column of type `gps` or `geopoint`.
Otherwise, return NA.
}
\examples{
good_eats_data <- onaDownload("good_eats", "mberg","mberg")
ge_spdf <- as.SpatialPointsDataFrame(good_eats_data)
class(ge_spdf) # "SpatialPointsDataFrame"
}
