% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona.R
\name{recastDataFrameBasedOnFormDF}
\alias{recastDataFrameBasedOnFormDF}
\title{Casts a dataframe to the right types based on a form-dataframe. Used by onaCast}
\usage{
recastDataFrameBasedOnFormDF(df, formdf, convert.dates = TRUE)
}
\arguments{
\item{df}{data}

\item{formdf}{form data frame. See format on onaCast.}
}
\value{
df re-casted data frame.
}
\description{
Casts a dataframe to the right types based on a form-dataframe. Used by onaCast
}
\examples{

#See examples under onaRead; this should be used through onaRead in almost all cases
}
