% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona.R
\name{addPhotoURLs}
\alias{addPhotoURLs}
\title{Add columns corresponding to the original, as well as medium and small thumbnails of images
as stored on the ona server.}
\usage{
addPhotoURLs(onaDataObj, onaUsername, type = "url")
}
\arguments{
\item{the}{onaData object which to create URLs from.}

\item{the}{ona username of the person who owns this form.}

\item{type:}{"url" or "img". URL only puts image url in, img puts image tag in.}
}
\value{
a onaData object, with a few additional URL columns
}
\description{
Add columns corresponding to the original, as well as medium and small thumbnails of images
as stored on the ona server.
}
\examples{
good_eats <- onaDownload("good_eats", "mberg","mberg")
good_eats_with_photos <- addPhotoURLs(good_eats, "mberg")
grep("URL", names(good_eats_with_photos), value=TRUE) # the new columns
}

