% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona.R
\name{onaDownload}
\alias{onaDownload}
\title{Download data from ona.}
\usage{
onaDownload(formName, account, uname, pass = NA, ...)
}
\arguments{
\item{formName}{formname on ona.io for which we download the data}

\item{account}{ona.io account/username/org_username that owns the form}

\item{uname}{ona.io username}

\item{pass}{ona.io password, if the data and/or form is private}

\item{...}{other parameters to pass onto onaRead}
}
\value{
onaDataObj a onaData Object, with "data" and "form" slots
}
\description{
This function downloads a dataset for the given form and username, and produces a 
onaData Object.
}
\examples{
good_eats <- onaDownload("good_eats", "mberg","mberg")
good_eats # is a data frame of all the data
good_eats@form # is the form for that data, encoded as a dataframe
privateData <- onaDownload("Private_Data_For_Testing", "ona_r", uname="ona_r", pass="ona_r")
}
