% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona.R
\name{onaRead}
\alias{onaRead}
\title{Reads data from a passed csv filename and json filename into a onaData object.}
\usage{
onaRead(csvfilename, jsonfilename, extraFormDF = data.frame(),
  dropCols = "", na.strings = c("n/a"), convert.dates = TRUE,
  keepGroupNames = TRUE)
}
\arguments{
\item{csvfilename}{filename (or a connection object) that has the ona data}

\item{jsonfilename}{filename of a json file (or a connection object) that has the form.json form}

\item{extraFormDF}{override the form (such as by providing a type for a calculate, a new label, etc.)}

\item{dropCols}{a regular expression, any column name that matches that regexp will be dropped}

\item{na.strings}{list of na.strings to be passed onto read.csv (default: "n/a")}

\item{keepGroupNames}{for a question with name foo in group bar, keepGroupName=T will generate
a name foo.bar, while keepGroupName=F will generate a name bar}
}
\value{
onaDataObj a onaData Object, with "data" and "form" slots
}
\description{
This function creates a onaData object from two files: a csv data file, and a 
json form file. These should both be downloaded from ona.io for the same form.
}

