\name{recastDataFrameBasedOnFormDF}
\alias{recastDataFrameBasedOnFormDF}
\title{Casts a dataframe to the right types based on a form-dataframe. Used by formhubCast}
\usage{
recastDataFrameBasedOnFormDF(df, formdf, convert.dates = TRUE)
}
\arguments{
  \item{df}{data}

  \item{formdf}{form data frame. See format on
  formhubCast.}
}
\value{
df re-casted data frame.
}
\description{
Casts a dataframe to the right types based on a
form-dataframe. Used by formhubCast
}
\examples{
#See examples under formhubRead; this should be used through formhubRead in almost all cases
}

