% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona.R
\name{replaceHeaderNamesWithLabels}
\alias{replaceHeaderNamesWithLabels}
\title{Get a new dataframe, where the header contains the full questions as opposed to slugs.}
\usage{
replaceHeaderNamesWithLabels(onaDataObj)
}
\arguments{
\item{onaDataObj}{is the ona data object whose data slot will be renamed}
}
\value{
a new data frame with the column names renamed from `name`s (slugs) to `label`s(full questions)
}
\description{
ona Objects have some data, as well as the form, which documents how
the data was obtained through a survey. The data, by default, is represented by
slugs, ie, items in the `name` column in the original xfrom. This function
replaces slugs in the header with the actual question text.
}
\examples{
good_eats <- onaDownload("good_eats", "mberg","mberg")
names(good_eats) # still slugged names
summary(good_eats$rating)
full_header_good_eats <- replaceHeaderNamesWithLabels(good_eats)
names(full_header_good_eats) # not slugged anymore
summary(full_header_good_eats$Rating) # but data is the same
}

